requirements.txt
fastapi
uvicorn
python-multipart
requests
jinja2
google-generativeai
Pillow
